Preset changed to: Patch
- changed name for main document in the database from preset to patch
- added README
